import of 0.2-rm+zomb-pre7
git-svn-id: http://svn.unix-ag.uni-kl.de/vpnc/trunk@8 315857ad-0bdb-0310-b42e-dec37551a5f0

+++ isakmp-pkt.c
@@ -266,7 +266,8 @@ void flatten_isakmp_packet (struct isakmp_packet *p,
     padding = blksz - (sz % blksz);
     if (padding == blksz)
       padding = 0;
-    DEBUG(3, printf("size = %d, blksz = %d, padding = %d\n", sz, blksz, padding));
+    DEBUG(3, printf("size = %ld, blksz = %ld, padding = %ld\n",
+	(long)sz, (long)blksz, (long)padding));
     flow_reserve (&f, padding);
   }
   f.base[lpos] = (f.end - f.base) >> 24;
